Created Loadout in basecreep.ts
