+++ package.json
@@ -13,5 +13,8 @@
     "@types/screeps-arena": "^0.0.1",
     "typedoc": "^0.22.15",
     "typescript": "^4.6.3"
+  },
+  "dependencies": {
+    "ts-pattern": "^4.0.1"
   }
 }
